# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "D&D 5E Character Creator API",
        "description": "RESTful API for creating and managing D&D 5th Edition (2024 ruleset) characters. Provides endpoints for character CRUD operations and reference data for classes, species, and backgrounds.",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/characters/{id}": {
            "get": {
                "tags": [
                    "Characters"
                ],
                "summary": "Get a character",
                "description": "Returns a single character with optional relationship includes.",
                "operationId": "cdbc78e4acce7e1799d4211bec16eab4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Character UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Comma-separated relationships to include",
                        "schema": {
                            "type": "string",
                            "example": "class,species,background,abilityScores"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Character"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Character retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Character not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Characters"
                ],
                "summary": "Delete a character",
                "description": "Soft deletes a character. The character is not permanently removed and can potentially be restored.",
                "operationId": "e7edf3b5ff708bd94cedffaa7f2928d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Character UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Character deleted successfully"
                    },
                    "404": {
                        "description": "Character not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Characters"
                ],
                "summary": "Update a character",
                "description": "Partially updates a character. Only include fields you want to change.",
                "operationId": "a6b0f235be26c5513e1f283c58836290",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Character UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCharacterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Character updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Character"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Character updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Character not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Character": {
                "required": [
                    "id",
                    "name",
                    "level"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    {
                        "properties": {
                            "id": {
                                "description": "Handles Character CRUD operations.\n\nAll business logic delegated to CharacterService.",
                                "type": "string",
                                "format": "uuid"
                            },
                            "name": {
                                "type": "string",
                                "example": "Gandalf the Grey"
                            },
                            "level": {
                                "type": "integer",
                                "example": 1
                            },
                            "experiencePoints": {
                                "type": "integer",
                                "example": 0
                            },
                            "maxHitPoints": {
                                "type": "integer",
                                "example": 10
                            },
                            "currentHitPoints": {
                                "type": "integer",
                                "example": 10
                            },
                            "temporaryHitPoints": {
                                "type": "integer",
                                "example": 0
                            },
                            "armorClass": {
                                "type": "integer",
                                "example": 10
                            },
                            "initiativeBonus": {
                                "type": "integer",
                                "example": 2
                            },
                            "speed": {
                                "type": "integer",
                                "example": 30
                            },
                            "proficiencyBonus": {
                                "type": "integer",
                                "example": 2
                            },
                            "inspiration": {
                                "type": "boolean",
                                "example": false
                            },
                            "alignment": {
                                "type": "string",
                                "enum": [
                                    "lawful_good",
                                    "neutral_good",
                                    "chaotic_good",
                                    "lawful_neutral",
                                    "true_neutral",
                                    "chaotic_neutral",
                                    "lawful_evil",
                                    "neutral_evil",
                                    "chaotic_evil"
                                ],
                                "nullable": true
                            },
                            "personalityTraits": {
                                "type": "string",
                                "nullable": true
                            },
                            "ideals": {
                                "type": "string",
                                "nullable": true
                            },
                            "bonds": {
                                "type": "string",
                                "nullable": true
                            },
                            "flaws": {
                                "type": "string",
                                "nullable": true
                            },
                            "backstory": {
                                "type": "string",
                                "nullable": true
                            },
                            "notes": {
                                "type": "string",
                                "nullable": true
                            },
                            "class": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/Class"
                                    }
                                ],
                                "nullable": true
                            },
                            "species": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/Species"
                                    }
                                ],
                                "nullable": true
                            },
                            "background": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/Background"
                                    }
                                ],
                                "nullable": true
                            },
                            "abilityScores": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/AbilityScores"
                                    }
                                ],
                                "nullable": true
                            },
                            "createdAt": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "ErrorResponse": {
                "required": [
                    "success",
                    "error",
                    "meta"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "required": [
                            "code",
                            "message"
                        ],
                        "properties": {
                            "code": {
                                "type": "string",
                                "example": "VALIDATION_ERROR"
                            },
                            "message": {
                                "type": "string",
                                "example": "The given data was invalid."
                            },
                            "details": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "timestamp": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "correlationId": {
                                "type": "string",
                                "format": "uuid"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}